split coin repository and asset repository
