+++ assets/data/src/main/java/com/easy/assets/data/di/AssetDataModule.kt
@@ -1,7 +1,9 @@
 package com.easy.assets.data.di
 
 import com.easy.assets.data.repository.AssetRepositoryImpl
+import com.easy.assets.data.repository.CoinRepositoryImpl
 import com.easy.assets.domain.repository.AssetRepository
+import com.easy.assets.domain.repository.CoinRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -15,4 +17,9 @@ abstract class AssetDataModule {
     abstract fun bindAssetRepository(
         impl: AssetRepositoryImpl
     ): AssetRepository
+
+    @Binds
+    abstract fun bindCoinRepository(
+        impl: CoinRepositoryImpl
+    ): CoinRepository
 }
