+++ assets/domain/src/main/java/com/easy/assets/domain/repository/AssetRepository.kt
@@ -1,12 +1,10 @@
 package com.easy.assets.domain.repository
 
-import com.easy.assets.domain.model.AssetInfo
 import com.easy.assets.domain.model.Transaction
 import com.easy.core.consts.ChainId
 import java.math.BigInteger
 
 interface AssetRepository {
-    fun address(slug: String, legacy: Boolean): String
     suspend fun signTransaction(slug: String): String
     suspend fun balance(address: String, chainId: ChainId, contractAddress: String?): BigInteger
     suspend fun transactions(
@@ -16,7 +14,4 @@ interface AssetRepository {
         limit: Int,
         contractAddress: String?
     ): Result<List<Transaction>>
-
-    suspend fun assets(): List<AssetInfo>
-    suspend fun assetsWithBalance(): Result<List<AssetInfo>>
 }
